US129047 - added sorting functionality back in
